Add FeeCalculationService

+++ src/EPAY.ETC.Core.API.Common/Models/TimeBlockFees/TimeBlockFeeModel.cs
@@ -10,9 +10,9 @@ namespace EPAY.ETC.Core.API.Core.Models.TimeBlockFees
         public Guid CustomVehicleTypeId { get; set; }
         [ForeignKey("CustomVehicletypeId")]
         public virtual CustomVehicleTypeModel? CustomVehicleType { get; set; }
-        public int FromSecond { get; set; } = 0;
-        public int ToSecond { get; set; } = 0;
+        public long FromSecond { get; set; } = 0;
+        public long ToSecond { get; set; } = 0;
         public int? BlockDurationInSeconds { get; set; } = 0;
-        public float? Amount { get; set; }
+        public double? Amount { get; set; }
     }
 }
